requirements.txt
absl-py==2.1.0
aiosmtplib==3.0.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
bcrypt==3.2.2
blinker==1.9.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
contourpy==1.3.0
cryptography==45.0.4
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.1
exceptiongroup==1.3.0
fastapi==0.115.1
fastapi-mail==1.5.0
flatbuffers==24.3.25
fonttools==4.55.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.3
mediapipe==0.10.18
ml_dtypes==0.5.0
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
passlib==1.7.4
pillow==11.0.0
protobuf==4.25.5
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
pytweening==1.2.0
PyYAML==6.0.2
reportlab==4.2.5
rsa==4.9.1
rubicon-objc==0.4.9
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.41
starlette==0.38.6
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zipp==3.21.0
